Print errors to stderr
Previously, errors would be dumped to stdout, which of course is not
ideal. Moreover, when the Arduino IDE runs arduino-builder --dump-prefs,
the errors would end up intermixed with the preferences, or be
completely ignored by the IDE when a fatal error occured. Printing
errors to stderr causes them to be printed properly by the IDE.

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>