Use absolute paths in #line directives generated
Multiple .ino files in a sketch are concatenated together, adding #line
directives so error messages refer to the original filenames. However,
these directives used plain filenames, without a path. Since these
filenames end up in the debug info as-is, this complicates using a
debugger on the resulting .elf file. Using full pathnames fixes this.

The tests are updated to expect full paths as well. This uses the
text/template package to process the sketch file, allowing it access to
the full context. This is a bit overkill, but it is easy and might be
useful for more complex testcases later.

This fixes arduino/Arduino#3746.

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>