Adding support for sketch-specific build properties
Tests for existence of a build_props.txt file in the sketch
directory. If present reads properties similar to those found in the
hardware and board files, allowing the user to easily customise the
build for a specific sketch. Most useful when containing lines similar
to

    compiler.c.extra_flags=-D NDEBUG
    compiler.cpp.extra_flags=-D NDEBUG -D MYLIBRARY_BUFSIZE=100

which enables the user to have macros defined for all compilation
units, including libraries.

Signed-off-by: Steve Marple <stevemarple@googlemail.com>