Be more conservative about what prototypes to add
Previously, a prototype was added if the generated prototype was
contained in the original source code line, to catch cases where ctags
did not provide enough information to generate a completely correct
prototype. However, this substring matching was still too liberal: if
something unsupported was present at the start of the function
definition (such as `static` in `static void func()`) the substring
matching would still think the prototype was correct, but the compiler
would complain.

This commit changes this to use prefix matching: Only if the original source
line starts with the generated prefix, a prototype is added. Ideally,
the full definition is matched, but that requires knowing where it ends,
which can be at a curly open, but also a comment start, etc., so better
to just leave it at this.

This means that inline functions and functions with attributes specified
at the start will no longer get a prototype, so the
TestPrototypesAdderSketchWithInlineFunction is modified accordingly.

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>