Added a build-all sketch test
This test tries to build all the test sketch proposed in issues.

A possible improvement is to find a way to easily add new sketch to the
test suite, possibily without the need to change the source code in

try_build_of_problematic_sketch_test.go

maybe a "go generate" approach may be investigated in this case.

Signed-off-by: Cristian Maglie <c.maglie@arduino.cc>