Moved Source and SourceGccMinusE into Context
Some helper "Command" now accepts directly the address of the target
string to work on instead of the key of the context map that contains
the string. Those "arguments" are, I guess, legacy from old
preprocessing and library discovery and IMHO they can be removed with a
bit of refactoring.

Anyway, I've intentionally limited the scope of this commit and left
further optimization for the future.

Signed-off-by: Cristian Maglie <c.maglie@arduino.cc>