Add utils.FindFilesInFolder
This is a new helper function that takes care of listing files with
given extensions in a directory, recursively or non-recursively. This
functionality was previously duplicated in a few places. This new helper
function now replaces some duplicate code in SketchLoader and
CollectAllSourcesFilesFromFoldersWithSources. Compilation of code in the
core and libraries still uses two other mechanisms for listing files (in
builder_utils), which will later be converted to use this new helper
function too.

This new helper function has some subtle differences compared to the
previous code:
 - The new code raises an error when the root directory does not exist,
   which the CollectAllSourcesFilesFromFoldersWithSources code did not
   do. To compensate for this, LibraryToSourceFolder will only return
   the "utility" folder when it exists.
 - The new code completely ignores any hidden or source control folders
   or files. Originally, arduino-builder showed a warning when source
   control or hidden folders were present, and then compiled files
   inside of them anyway. Since f5edd23 (Avoid warning about SCCS folders
   in libraries, just ignore them) this warning is no longer shown in
   most cases, but the files themselves are still included in
   compilation. With this commit, they are just completely ignored
   instead (for sketch compilation and include detection, for core and
   library compilation they will are still used). See also #139.

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>