Fix duplicate-skipping in prototype generation
While generating prototypes, a scan is made for any function definitions
which would result in identical prototypes, and all but the first are
skipped. However, when doing so, this did not take into account function
definitions which were already marked to be skipped, which could cause
*all* definitions to be skipped.

This behaviour caused a particular problem when a method definition
existed with the same signature as a function definition. All method
definitions are skipped through `tagIsUnhandled`, but could still
make `skipDuplicates` skip the corresponding function definition. This
commit fixes this.

It is not entirely clear to me if this duplicate removal is actually
needed at all. If code contains multiple function definitions with the
same signature, you would expect the compiler to throw an error anyway.
If duplicate removal would be removed, only one testcase
(TestCTagsToPrototypesShouldDealFunctionWithDifferentSignatures) fails,
but that testcase uses a predefined ctags output that contains a
duplicate function definition. It is not quite clear what source files
this output was generated from (it seems to stem from
CharWithEscapedDoubleQuote.ino, which originated at
https://github.com/arduino/Arduino/issues/1245, but those only contain
one version of the function definition, not two).

Still, since the ctags parsing will hopefully be replaced in the near
future, this just fixes the duplicate removal instead of removing it and
risking a regression on some corner case.

This fixes #140.

Signed-off-by: Martino Facchin <m.facchin@arduino.cc>
Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>