Fix tool loading testcases
In commit 7520e66 (Removed no more used "coan" tools), one tool was
removed, but these testcases were not updated accordingly. This commit
updates the testcases with the correct number of tools to expect.

While here, the testcases are expanded to also explicitely check the name
and version number for all tools. This makes the test more complete, but
also documents what tools and versions are expected exactly, which makes
it easier to diagnose problems like this one later.

Also, instead of using hardcoded array indices, this converts the latter
two testcases to use an idx variable, just like the first testcase. This
makes it easier to insert a new tool later.

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>