Limit recursive sketch compilation to the `src` directory
Since c7a18f8 (Sketch sources files were recursively copied BUT not
compiled), sketches were compiled recursively. However, this turned out
to have some side effects, such as including files that were not
intended to be compiled (included as documentation or otherwise not part
uof the sketch itself) and breaking the build if the build directory is
inside the sketch directory (#89).

To prevent these side effects, but still allow users to apply more
structure to their sketches, recursive compilation is now limited to the
`src` directory inside the sketch folder. All .ino files must still
reside in the root folder, but any .c and .cpp files inside
(subdirectories of) the `src` directory are included in the build.

Using a `src` directory in this way is similar to how libraries are
compiled (except for libraries *all* source files live in the `src`
directory if it is present, while for sketches files can *also* live in
the root directory).

Sketches written to use this new `src` directory should also be backward
compatible, in the sense that arduino-builder 1.0.7 (Arduino IDE 1.6.7)
and later can succesfully compile it, not just the versions after this
commit. The reverse is not true: sketches that need recursive
compilation outside of the `src` directory, as supported by
arduino-builder 1.0.7 above, will not compile after this commit, but it
is trivial to convert them to do so (without breaking compilation before
this commit).

This also changes some testcases that used subfolders, to move these
inside the `src` directory instead of directly in the sketch root
directory.

This fixes #89.

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>