Revert fix for #37 and always show "multiple libraries warning" again
This reverts commit 20584795c86c40be9073ded245264f7a6becf881:

	Adding info in LibraryResolutionResult about whether selected
	library comes from a platform or from outside. If the former and
	there are duplicates, no warning is printed

Additionally, the librariesInSomePlatform function and
librariesInPlatforms variable in resolveLibrary are removed, since these
are now unused.

The original suggestion in #37 was that if a platform library overrides
a builtin library, no warning would be needed. The implementation,
however, was hiding the warning whenever a platform library was
overriding any other library, including a user library, which seems
harmful (This was [reported on the devlist][1]).

This commit reverts the exception added for #37, making the warning
appear unconditionally again. In a future commit, a better solution
for #37 will be added.

[1]: https://groups.google.com/a/arduino.cc/d/msg/developers/1kkIqIsbuzU/0-abwr1gBQAJ

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>