Properly escape filenames in #line directives
In some cases, backslashes in a filenames were already escaped when
emitting a #line directive, but sometimes no escaping happened. In any
case, quotes in a filename should also be escaped.

This commit introduces a helper function `QuoteCppString()` to handle
quoting and escaping strings in a generic way. This new helper function
is used in the code itself, but not in testcases yet.

Note that ctags does not quite properly handle special characters and
escaping in filenames. When parsing line markers, it just ignores
escapes, so any escapes from the #line marker will be present in the
ctags output as well. Before this commit, they would be copied
unmodified into the #line directives generated from the ctags filenames.  Now that these line directives have escaping applied, the filenames read from ctags need to be unescaped, to prevent double escapes. This
unescaping already happened in CollectCTagsFromSketchFiles, but is now
moved to parseTag where it is more appropriate and applies to all users.

Since ctags does not handle escapes in line markers, it cuts the
filename off at the first double quote. If a filename contains a double
quote, the filename as output by ctags will be cut off before it,
leaving the escaping \ at the end. Before this commit, this trailing \
would be output directly into the generated #line marker, which confused
gcc and made the build fail. With this commit, this trailing \ is
escaped itself, making the output syntactically valid, so the build now
works with double quotes in a filename. However, due to this filename
cut off, arduino-builder will no longer generate prototypes for
functions in files that have a double quote in them (since it can no
longer detect them as sketch functions), so double quotes (and probably
tabs too) in filenames are still not completely supported yet.

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>