Use QuoteCppString in testcases
In some testcases, the generated output is checked, which contains #line
directives with filenames. Previously, this compared against the
filename with just backslashes escaped, but this would be incorrect when
there are quotes in the filename. Since the previous commit, the
arduino-builder code itself uses a new `utils.QuoteCppString` function to
apply escaping, so this applies that same escaping.

Part of the changes modify the test code directly, part of the changes
replace the `EscapeBackSlashes` "function" that was used in
interpolating some text files with a call to QuoteCppString instead.

Note that QuoteCppString already adds quotes around its result, so those
are removed from the places where this function is called.

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>