Let IncludesFinderWithRegExp return at most 1 include
Originally, this code was written to find all include statements in a
source file, for further processing. However, since a while the
include finding was changed to run the preprocessor to see if any
includes are missing, and this class is used to find the missing
includes in the error message.

In practice, the preprocessor will bail out at the first missing
include, since there is no meaningful way for it to continue. This means
there will be at most one include in the error output. If there
would be multiple includes (either because the preprocessor decided to
continue, or there is some other output that accidentially matches the
regex), it would be harmful to actually process all of them, since the
absence of the first include might influence the presence of the second
(e.g. in the presence of #ifdefs).

This commit enforces that only one include is returned at a time,
slightly simplifying the code in the process. The filename and class
should be renamed to singular as well, but this will be left for a
future commit to reduce noise.

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>