Do not limit includes to the `src` sketch subdirectory
Commit 405a24a (Limit recursive sketch compilation to the `src`
directory) modified the sketch loading to only recursively load files
from the `src` subdirectory of a sketch, in order to only compile source
files in that subdirectory. As a side effect, this also prevents source
and header files in other subdirectories from being copied into the
build directory, and hence from being included. This side effect was not
entirely unanticipated, but its implications were underestimated.

This commit undoes these side effects: source and header files are again
loaded (and thus copied) recursively in all subdirectories of a sketch.
Compilation is still limited to the src subdirectory (there already was
code to achieve this). This commit also modifies include detection to be
limited to the src subdirectory (which was previously implicit, since it
can only look at files that were copied into the build directory).

With this commit, the the meaning of "loading" a file becomes a bit
fuzzy in the code, but this is intended to be fixed in a future
refactor.

A test will be added in the next commit.

This fixes arduino/Arduino#5186.

Signed-off-by: Cristian Maglie <c.maglie@arduino.cc>
Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>