Remove unneeded code from ContainerFindIncludes
The removed code iterates over the detected libraries and adds their
source directories to the queue for running further detection on them.
However, IncludesToIncludeFolders (called by findIncludesUntilDone(),
which is called a bit above the removed code) already does this, so
there is no need to do it again. Since Ctx.FoldersWithSources is a
UniqueStringQueue, this unneeded code was not affecting behaviour in any
way.

It seems likely that this code had a purpose previously, but it was lost
in some refactor. However, a quick scan of the commit that introduced it
(d6e378: Implemented library to library dependency full discovery)
suggests that this code was useless from the start.

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>