Call findIncludesUntilDone from only one place
Previously, it was called once for the main cpp file, and then in a loop
for any extra source files from the sketch or dependent libraries.
However, this can be simplified by just adding the main .cpp file to the
queue, and then letting the loop handle it.

This produces a tiny change in the order in which source files are
processed for includes, but any setups affected by this are already
prone to errors.

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>