Remove Context.FoldersWithSourceFiles
This field was used during include detection, to collect the source
folders whose source files should be processed for includes. This field
was handled by the CollectAllSourceFilesFromFoldersWithSources pass,
which finds all source files in these folders and adds them to the
Context.CollectedSourceFiles queue instead.

This commit changes all code that used to add directories to the
FoldersWithSourceFiles queue to instead add all contained source files
to the CollectedSourceFiles queue. For this, the
CollectAllSourceFilesFromFoldersWithSources pass is transformed into
regular function and renamed to QueueSourceFilesInFolder, so it can be
directly called without needing any particular context entries.

Note that the filename that contains the latter function is no longer
entirely accurate, but this will be fixed in a later commit (left
untouched in this commit to simplify review).

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>