Remove IncludesToIncludeFolders.Run
In previous commits, this method has been signficantly simplified,
reducing it to little more than a wrapper around the resolveLibrary
helper function.

This commit renames resolveLibrary to ResolveLibrary to make it public
and lets findIncludesUntilDone call it directly (instead of going
through IncludesToIncludeFolders). findIncludesUntilDone now also takes
care of processing the ResolveLibrary result, which was previously done
by IncludesToIncludeFolders.

The signature of ResolveLibrary is also changed to accept a Context, so
it can now extract the needed variables itself, instead of needing
IncludesToIncludeFolders to extract them.

Note that the filename that contains ResolveLibrary is no longer
entirely accurate, but this will be fixed in a later commit (to simplify
review of this commit).

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>