Add SourceFile struct
This struct contains the path of a source file, together with the
library or sketch it is contained in. This allows the SourceFile methods
to generate the full paths to the source file, object file and
dependency file, making it easier to pass around source files.

This commit only adds the struct. Next commits will start to use this
struct in the include detection, where it fills an immediate need, later
it will be used in more places as well.

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>