Store SourceFiles in Context.CollectedSourceFiles
Previously, the include detection kept a queue of strings representing
full paths to source files that still needed to be processed and
iterated over it. This commit turns this into a queue containing the
(new) SourceFile struct. This gives the include detection code a bit
more context about a source file, allowing to also generate the object
and dependency file paths.

Previously this was not feasible, since it did not know the origin
(library/sketch) of a path, so it could not decide where within the
build path the object file should live.

This does not actually use this new context yet, but it will be useful
when introducing caching next.

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>