Move QueueSourceFilesFromFolder into container_find_includes.go
This function used to be a pass, but it was reduced to a regular
function in 1001916 (Remove Context.FoldersWithSourceFiles). By now, the
function is only called from container_find_includes.go, so it makes
sense to just move it into that file. This does not change any code, it
just moves the function and renames it to be private.

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>