Show errors during include detection properly
When an include could not be resolved, the preprocessor was ran again to
show the include error as reported by the compiler to the user. However,
the preprocessor was now always run against the main sketch, instead of
the file currently being processed. Unless the problem was in fact in
the main sketch, this caused no error to be shown and include detection
continued with the next file instead of being aborted as intended.

This fixes the problem by running the preprocessor on the right file. In
the future, this error handling might be a bit more improved, but for
now this should be sufficient.

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>