Put utility folders in the include path during include detection
When compiling source files from a library, the utility folder of the
library itself is in the include path. When doing include detection,
this did not happen, which could lead to compilation errors. Because of
the broken (but recently fixed) errorhandling in include detection, this
error was not visible to the user. It would only have a visible effect
if a library included another library *after* including a file from its
utility folder through the include path, and that other library was not
otherwise pulled in.

In practice, this would almost never occur, but with the new caching the
cache would be invalidated and this problem became visible.

This is fixed by simply including the utility folder in the include path
during include detection. This required explicitly passing the include
path through the GCCPreprocRunnerForDiscoveringIncludes and
GCCPreprocRunner.

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>