CTagsParser is now a standard go class
This is a first refactoring that aims to create a ctags package that
is not strongly tied to the builder (and so can possibly be replaced
with another parser in the future)

A temporary CTags command has been made to reduce the number of changes
involved in this commit. It will be removed at the end of the
refactoring.

Signed-off-by: Cristian Maglie <c.maglie@arduino.cc>