Generalize ObjFileIsUpToDate to BuildResultIsUpToDate
This renames this function and adds a resultFile parameter. This
resultFile is used for all timestamp checks against other files (source
file, dep file, header files from the dep file), while the objectFile
passed is now only used to check against the contents of the dep file.

Both calls to this function still pass the same filename as objectFile
and resultFile, so this commit should not change any behavior.

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>