Allow precompiled objects to be included in the build
Combined with Library.precompiled flag, allows a developer to ship a precompiled library.

The linker (ld in our case) needs static libraries to be explicitely included in the build (it doesn't try any heuristic).
Luckily, "including" a dynamic library in this way works as well, so we can avoid the "-L" and "-l" dance