Improve build instructions in README
This drops the `setup_go_env_vars` script, which is only confusing. It
contained three lines, of which only one (setting `GOPATH`) is always
required and the other two are only relevant when go is manually
installed. In the most common case (go installed through a package
manager), they will even break the build.

Instead of referring to an external script, just show how to set
`GOPATH` in the README itself.

This improvement was prompted by the user confusion found in #244.

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>