Print "archiving core" message at INFO level
This previously used the DEBUG level, which is shown in red by the
Arduino IDE. Commit 7883e01 (Let "archiving core" message appear only if
verbose) attempted to fix this, but did not.

While here, this also changes Fprintln with an explicit Stdout argument
to a regular Println, since Stdout is the default anyway.

This commit properly fixes #238.

Signed-off-by: Matthijs Kooijman <matthijs@stdin.nl>