Introduce compiler.libraries.ldflags
Choosing compiler.ldflags for precompiled libraries was an unfortunate choice, since some cores define it in "wrong" places (or it's already in use for other flags).

Deprecate the old behaviour and add a specific flag that should be added to recipe.combine in the right position to let the linker find the libraries.