Don't bail out if got include unrelated error while preprocessing
Preprocessing is a bit tricky, since it expects an error to resolve missing includes; to do this, it invokes gcc with -CC flag (https://gcc.gnu.org/onlinedocs/gcc-4.4.0/gcc/Preprocessor-Options.html#Preprocessor-Options) to expand macro AND comments. Since this flag is not used during normal compilation, the errors reported are mostly bogus.

This patch avoids writing the actual (non-include) error on stdout so the compilation can go on and fail on the proper errors (if there's any)