cmake: Allow dynamic linking of libraries
This commit introduce a way for library authors to force the builder to
output a cmake build script with a library linked dynamically.

The trick is to add a file named arduino_builder.properties with the
following key:

   cmake.pkg_config=lib1 lib2 lib3 ...

By doing so, the lib archives (.a) in the library are ignored and a
bunch of "pkg_search_module" rows are added to the CMakeList.txt for
the specified "lib1" "lib2" "lib3".
The corresponding flags in "link_directories" and "target_link_libraries"
directives are added as well.

The above trick does affect only the cmake export, the normal build of
the arduino-builder is unchanged.