Don't scan hidden directories (.git, etc)
When checking for updated core files, don't scan inside UNIX hidden
directories like .git.  Scanning inside a GIT tree for a large project
with many branches could take many minutes of runtime and multiple GB of
RAM.  This was seen in the Arduino core for the ESP8266, but is probably
also an issue for other core development teams.

Fixes #327

Signed-off-by: Earle F. Philhower, III <earlephilhower@yahoo.com>